+++ queen_store/src/main/java/com/example/queen_store/common/order/Order.java
@@ -1,76 +1,4 @@
 package com.example.queen_store.common.order;
 
-import java.util.Date;
-import java.util.Objects;
-
 public class Order {
-    private int orderId;
-    private Date orderDate;
-    private double totalPrice;
-    private int userId;
-    private int voucherId;
-
-    public Order() {
-    }
-
-    public Order(int orderId, Date orderDate, double totalPrice, int userId, int voucherId) {
-        this.orderId = orderId;
-        this.orderDate = orderDate;
-        this.totalPrice = totalPrice;
-        this.userId = userId;
-        this.voucherId = voucherId;
-    }
-
-    public int getOrderId() {
-        return orderId;
-    }
-
-    public void setOrderId(int orderId) {
-        this.orderId = orderId;
-    }
-
-    public Date getOrderDate() {
-        return orderDate;
-    }
-
-    public void setOrderDate(Date orderDate) {
-        this.orderDate = orderDate;
-    }
-
-    public double getTotalPrice() {
-        return totalPrice;
-    }
-
-    public void setTotalPrice(double totalPrice) {
-        this.totalPrice = totalPrice;
-    }
-
-    public int getUserId() {
-        return userId;
-    }
-
-    public void setUserId(int userId) {
-        this.userId = userId;
-    }
-
-    public int getVoucherId() {
-        return voucherId;
-    }
-
-    public void setVoucherId(int voucherId) {
-        this.voucherId = voucherId;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Order order = (Order) o;
-        return orderId == order.orderId;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(orderId);
-    }
 }
